now that the types have been converted, revert some that shouldn't have been

+++ macosx/FileOutlineView.m
@@ -176,7 +176,7 @@
     }
 }
 
-- (void) drawRow: (int) row clipRect: (NSRect) clipRect
+- (void) drawRow: (NSInteger) row clipRect: (NSRect) clipRect
 {
     if (![self isRowSelected: row])
     {
